chore: move fields group into render
