+++ documents/passport_bottom.py
@@ -6,9 +6,8 @@ from PIL.Image import Resampling
 
 from make_dataset import DocumentGenerator
 from make_dataset.factories.fields import FieldFactory
-from make_dataset.fields import FieldGroup
 from make_dataset.positioning import Point
-from make_dataset.render import Font, renderers
+from make_dataset.render import FieldGroup, Font, renderers
 
 faker = Faker('ru_RU')
